pyproject.toml
[tool.poetry]
name = "cource-moniroring-notifier"
version = "0.1.0"
description = ""
authors = ["Mimrik <misha@mamrov.ru>"]
readme = "README.md"
source = [
    {name = "PyPI", priority = "default"},
    {name = "abm_pypi", url = "https://pypi2.abm-jsc.ru", priority = "primary"},
]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2"
pytz = "^2024.1"
init-helpers-abm={version="1.2.59931", source="abm_pypi"}
async-tools-abm={version="2.1.57881", source="abm_pypi"}
sqlalchemy-tools-abm={version="2.0.58719", source="abm_pypi"}
http-tools-abm={version="5.3.59932", source="abm_pypi"}
coverage = "^7.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
